# pkg/http/static/assets/docs/b3scale-openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "b3scale api v1",
    "version": "1.1.0",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This document describes the specifications for the b3scale API v1.",
    "contact": {
      "name": "The B3Scale Developers",
      "url": "https://b3scale.io",
      "email": "mail@infra.run"
    }
  },
  "paths": {
...
    "/v1/recordings": {
      "get": {
        "description": "Retrieve the recordings collection. Mandatory filter is by either `frontend_key` or `frontend_id`.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Recordings"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/InvalidJWTError"
          }
        },
        "operationId": "recordingsList",
        "parameters": [
          {
            "description": "Filter recordings by frontend-key.",
            "in": "query",
            "name": "frontend_key",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter recordings by frontend-id.",
            "in": "query",
            "name": "frontend_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "List recordings",
        "tags": [
          "Recordings"
        ]
      }
    },
...
    "/v1/recordings/{id}": {
      "get": {
        "description": "Fetch a single recording identified by ID.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Recording"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/InvalidJWTError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        },
        "operationId": "recordingsRead",
        "summary": "Read",
        "tags": [
          "Recordings"
        ]
      },
      "parameters": [
        {
          "description": "The identifier of the object.",
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "description": "A general error response object.",
        "properties": {
          "error": {
            "type": "string",
            "description": "An error type tag. See specific error for details."
          },
          "message": {
            "type": "string",
            "description": "A human readable message with details about the error."
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
      "Format": {
        "description": "Format",
        "properties": {
          "Length": {
            "type": "integer"
          },
          "Preview": {
            "$ref": "#/components/schemas/Preview"
          },
          "ProcessingTime": {
            "type": "integer"
          },
          "Type": {
            "type": "string"
          },
          "URL": {
            "type": "string"
          }
        },
        "required": [
          "Type",
          "URL",
          "ProcessingTime",
          "Length",
          "Preview"
        ],
        "type": "object"
      },
...
      "Image": {
        "description": "Image",
        "properties": {
          "All": {
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "type": "array"
          }
        },
        "required": [
          "Alt",
          "Height",
          "Width",
          "URL"
        ],
        "type": "object"
      },
      "Images": {
        "description": "Images",
        "properties": {
          "All": {
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "type": "array"
          }
        },
        "required": [
          "All"
        ],
        "type": "object"
      },
...
      "NotFoundError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          }
        ],
        "description": "The requested resource could not be found",
        "type": "object"
      },
      "Preview": {
        "description": "Preview",
        "properties": {
          "Images": {
            "$ref": "#/components/schemas/Images"
          }
        },
        "required": [
          "Images"
        ],
        "type": "object"
      },
...
      "Recording": {
        "description": "Recording",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "frontend_id": {
            "description": "The id of the associated frontend.",
            "type": "string"
          },
          "internal_meeting_id": {
            "description": "The internal meeting id.",
            "type": "string"
          },
          "meeting_id": {
            "description": "The id of the related meeting.",
            "type": "string"
          },
          "record_id": {
            "description": "ID of the recording.",
            "type": "string"
          },
          "recording": {
            "$ref": "#/components/schemas/RecordingData",
            "description": "The bbb recording data"
          },
          "synced_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "record_id",
          "recording",
          "meeting_id",
          "internal_meeting_id",
          "frontend_id",
          "created_at",
          "updated_at",
          "synced_at"
        ],
        "type": "object"
      },
      "RecordingData": {
        "description": "Recording",
        "properties": {
          "EndTime": {
            "type": "integer"
          },
          "Formats": {
            "items": {
              "$ref": "#/components/schemas/Format"
            },
            "type": "array"
          },
          "InternalMeetingID": {
            "type": "string"
          },
          "IsBreakout": {
            "type": "boolean"
          },
          "MeetingID": {
            "type": "string"
          },
          "Metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "Name": {
            "type": "string"
          },
          "Participants": {
            "type": "integer"
          },
          "Published": {
            "type": "boolean"
          },
          "RecordID": {
            "type": "string"
          },
          "StartTime": {
            "type": "integer"
          },
          "State": {
            "type": "string"
          }
        },
        "required": [
          "RecordID",
          "MeetingID",
          "InternalMeetingID",
          "Name",
          "IsBreakout",
          "Published",
          "State",
          "StartTime",
          "EndTime",
          "Participants",
          "Metadata",
          "Formats"
        ],
        "type": "object"
      },
...
      "Recordings": {
        "description": "List of Recordings",
        "items": {
          "$ref": "#/components/schemas/Recording"
        },
        "type": "array"
      },
...
      "ServerError": {
        "description": "A Server Error",
        "properties": {
          "message": {
            "type": "string",
            "description": "A human readable message with details about the error."
          }
        },
        "type": "object"
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "The request was invalid. Maybe a JWT authorization header was not provided or a validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ServerError"
            }
          }
        }
      },
...
      "InvalidJWTError": {
        "description": "The JWT authorization token was invalid or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ServerError"
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "The requested resource could not be found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NotFoundError"
            }
          }
        }
      },
...
      "Recording": {
        "description": "Recording",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Recording"
            }
          }
        }
      },
...
      "Recordings": {
        "description": "List of Recordings",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Recordings"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "description": "Authorization using a JWT bearer token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "default server location"
    }
  ],
  "security": [
    {
      "jwt": []
    }
  ]
}